refactor(p2p_network/client): move  p2p types to the p2p crate
