+++ crates/p2p/src/client.rs
@@ -1,2 +1,3 @@
 pub mod peer_agnostic;
 pub mod peer_aware;
+pub mod types;
