+++ crates/pathfinder/src/p2p_network/client/v1.rs
@@ -5,20 +5,17 @@ pub mod types {
     };
     use pathfinder_common::{
         event::Event,
-        state_update::SystemContractUpdate,
         transaction::{
             DeclareTransactionV0V1, DeclareTransactionV2, DeployAccountTransaction,
             DeployTransaction, EntryPointType, InvokeTransactionV0, InvokeTransactionV1,
             L1HandlerTransaction, TransactionVariant,
         },
-        BlockHash, BlockNumber, BlockTimestamp, CallParam, CasmHash, ClassHash, ConstructorParam,
-        ContractAddress, ContractAddressSalt, ContractNonce, EntryPoint, EthereumAddress,
-        EventData, EventKey, Fee, GasPrice, L1ToL2MessageNonce, L1ToL2MessagePayloadElem,
-        L2ToL1MessagePayloadElem, SequencerAddress, StarknetVersion, StorageAddress, StorageValue,
-        TransactionNonce, TransactionSignatureElem, TransactionVersion,
+        CallParam, CasmHash, ClassHash, ConstructorParam, ContractAddress, ContractAddressSalt,
+        EntryPoint, EthereumAddress, EventData, EventKey, Fee, L1ToL2MessageNonce,
+        L1ToL2MessagePayloadElem, L2ToL1MessagePayloadElem, TransactionNonce,
+        TransactionSignatureElem, TransactionVersion,
     };
     use starknet_gateway_types::reply::transaction as gw;
-    use std::{collections::HashMap, time::SystemTime};
 
     /// We don't want to introduce circular dependencies between crates
     /// so in those cases we cannot use TryFrom and we need to work around for the orphan rule
@@ -29,48 +26,9 @@ pub mod types {
             Self: Sized;
     }
 
-    /// Simple block header meant for the temporary p2p client hidden behind
-    /// the gateway client api, ie.: does not contain any commitments
-    ///
-    /// TODO: remove this once proper p2p friendly sync is implemented
-    #[derive(Debug, Clone, PartialEq)]
-    pub struct BlockHeader {
-        pub hash: BlockHash,
-        pub parent_hash: BlockHash,
-        pub number: BlockNumber,
-        pub timestamp: BlockTimestamp,
-        pub gas_price: GasPrice,
-        pub sequencer_address: SequencerAddress,
-        pub starknet_version: StarknetVersion,
-    }
-
-    /// Simple state update meant for the temporary p2p client hidden behind
-    /// the gateway client api, ie.:
-    /// - does not contain any commitments
-    /// - does not specify if the class was declared or replaced
-    ///
-    /// TODO: remove this once proper p2p friendly sync is implemented
-    ///
-    /// How to manage this modest state update:
-    /// 1. iterate through contact updates and check in the db if the contract is already there to figure out
-    ///    if it means replacement or declaration
-    /// 2. take the remaining ones which are then treated as declared and then figure out which is Cairo 0 and which is Sierra
-    #[derive(Default, Debug, Clone, PartialEq)]
-    pub struct StateUpdate {
-        pub contract_updates: HashMap<ContractAddress, ContractUpdate>,
-        pub system_contract_updates: HashMap<ContractAddress, SystemContractUpdate>,
-    }
-
-    #[derive(Default, Debug, Clone, PartialEq)]
-    pub struct ContractUpdate {
-        pub storage: HashMap<StorageAddress, StorageValue>,
-        /// The class associated with this update as the result of either a deploy or class replacement transaction.
-        /// We don't explicitly know if it's one or the other
-        pub class: Option<ClassHash>,
-        pub nonce: Option<ContractNonce>,
-    }
-
     /// Represents a simplified receipt (events and execution status excluded).
+    ///
+    /// This type is not in the `p2p` to avoid `p2p` dependence on `starknet_gateway_types`.
     #[derive(Clone, Debug, PartialEq)]
     pub struct Receipt {
         pub actual_fee: Fee,
@@ -90,103 +48,6 @@ pub mod types {
                 l2_to_l1_messages: r.l2_to_l1_messages,
                 revert_error: r.revert_error.unwrap_or_default(),
             }
-        }
-    }
-
-    impl From<pathfinder_common::BlockHeader> for BlockHeader {
-        fn from(h: pathfinder_common::BlockHeader) -> Self {
-            Self {
-                hash: h.hash,
-                parent_hash: h.parent_hash,
-                number: h.number,
-                timestamp: h.timestamp,
-                gas_price: h.gas_price,
-                sequencer_address: h.sequencer_address,
-                starknet_version: h.starknet_version,
-            }
-        }
-    }
-
-    impl From<pathfinder_common::StateUpdate> for StateUpdate {
-        fn from(s: pathfinder_common::StateUpdate) -> Self {
-            Self {
-                contract_updates: s
-                    .contract_updates
-                    .into_iter()
-                    .map(|(k, v)| (k, v.into()))
-                    .collect(),
-                system_contract_updates: s.system_contract_updates,
-            }
-        }
-    }
-
-    impl From<pathfinder_common::state_update::ContractUpdate> for ContractUpdate {
-        fn from(c: pathfinder_common::state_update::ContractUpdate) -> Self {
-            Self {
-                storage: c.storage,
-                class: c.class.map(|x| x.class_hash()),
-                nonce: c.nonce,
-            }
-        }
-    }
-
-    impl TryFrom<p2p_proto_v1::block::BlockHeader> for BlockHeader {
-        type Error = anyhow::Error;
-
-        fn try_from(proto: p2p_proto_v1::block::BlockHeader) -> anyhow::Result<Self> {
-            Ok(Self {
-                hash: BlockHash(proto.hash.0),
-                parent_hash: BlockHash(proto.parent_hash.0),
-                number: BlockNumber::new(proto.number)
-                    .ok_or(anyhow::anyhow!("Invalid block number > i64::MAX"))?,
-                timestamp: BlockTimestamp::new(
-                    proto.time.duration_since(SystemTime::UNIX_EPOCH)?.as_secs(),
-                )
-                .ok_or(anyhow::anyhow!("Invalid block timestamp"))?,
-                gas_price: GasPrice::from_be_slice(proto.gas_price.as_slice())?,
-                sequencer_address: SequencerAddress(proto.sequencer_address.0),
-                starknet_version: StarknetVersion::from(proto.starknet_version),
-            })
-        }
-    }
-
-    // FIXME add missing stuff to the proto representation
-    impl TryFrom<p2p_proto_v1::state::StateDiff> for StateUpdate {
-        type Error = anyhow::Error;
-
-        fn try_from(proto: p2p_proto_v1::state::StateDiff) -> anyhow::Result<Self> {
-            const SYSTEM_CONTRACT: ContractAddress = ContractAddress::ONE;
-            let mut system_contract_update = SystemContractUpdate {
-                storage: Default::default(),
-            };
-            let mut contract_updates = HashMap::new();
-            proto.contract_diffs.into_iter().for_each(|diff| {
-                if diff.address.0 == SYSTEM_CONTRACT.0 {
-                    diff.values.into_iter().for_each(|x| {
-                        system_contract_update
-                            .storage
-                            .insert(StorageAddress(x.key), StorageValue(x.value));
-                    });
-                } else {
-                    contract_updates.insert(
-                        ContractAddress(diff.address.0),
-                        ContractUpdate {
-                            storage: diff
-                                .values
-                                .into_iter()
-                                .map(|x| (StorageAddress(x.key), StorageValue(x.value)))
-                                .collect(),
-                            class: diff.class_hash.map(ClassHash),
-                            nonce: diff.nonce.map(ContractNonce),
-                        },
-                    );
-                }
-            });
-
-            Ok(Self {
-                contract_updates,
-                system_contract_updates: [(SYSTEM_CONTRACT, system_contract_update)].into(),
-            })
         }
     }
